I refactored configure quite a bit. Most of this is to treat IP Pascal on equal footing with GPC.

+++ configure.bat
@@ -19,16 +19,16 @@ rem Determine if needed programs exist. The only fatal one is grep, because we
 rem need that to run this script. The rest will impact the running of various
 rem test and build scripts.
 rem
+
 where /q grep
-if %errorlevel% equ 0 goto grepexists
-echo *** No grep was found
-goto stop
-:grepexists
-where /q gpc
-if %errorlevel% equ 0 goto gpcexists
-echo *** No gpc was found
-goto stop
-:gpcexists
+if %errorlevel% neq 0 (
+
+    echo *** No grep was found
+    echo Terminating
+    exit /b 1
+    
+)
+
 where /q diff || echo *** No diff was found
 where /q sed || echo *** No sed was found
 where /q rm || echo *** No rm was found
@@ -40,31 +40,76 @@ where /q gzip || echo *** No zip was found
 where /q tar || echo *** No tar was found
 
 rem
-rem Evaluate compiler version and word size
+rem Check flip (Unix/dos end of line changer) exists, and make it if not.
 rem
-gpc -v 2> temp
-grep "gpc version 20070904" temp > temp2
-if %errorlevel% equ 0 goto gpccorrectver
-echo *** Warning, Pascal-P5 is only validated to work with gpc version 20070904
-:gpccorrectver
-rm temp2
-set bits=64
-grep "build=x86_64" temp > temp2
-if %errorlevel% equ 0 goto gpc64
-set bits=32
-
 where /q flip
-if %errorlevel% equ 0 goto flipexists
-echo Making flip.exe
-call bin\make_flip
-dir bin\flip.exe > temp
-grep "flip.exe" temp > temp2
-if %errorlevel% equ 0 goto flipexists
-echo *** Unable to make flip
-:flipexists
+if %errorlevel% neq 0 (
+
+    echo Making flip.exe
+    call bin\make_flip
+    dir bin\flip.exe > temp
+    grep "flip.exe" temp > temp2
+    if %errorlevel% neq 0 (
+    
+        echo *** Unable to make flip
+        
+    )
+    
+)
 
 rem
-rem Check all arguments
+rem Detect and prioritize IP Pascal. The reason for this is, if IP Pascal
+rem exists on the system, that is the preferable compiler.
+rem pc.exe is the IP Pascal command shell. You might have to fix this up
+rem if there is another pc.exe in the system.
+rem
+where /q pc
+if %errorlevel% equ 0 (
+
+    set compiler=ip_pascal
+    
+) else (
+
+    rem
+    rem Now check for GPC. Output scary message for no compiler found, but 
+    rem otherwise do nothing. rem Its up to the user to find a compiler.
+    rem 
+    where /q gpc
+    if %errorlevel% neq (
+
+        echo *** No gpc or pc was found, there is no ISO 7185 Pascal compiler installed
+    
+    ) else (
+    
+        rem
+        rem Evaluate GPC compiler version and word size. This is required with GPC
+        rem because not all versions work as ISO 7185 Compilers. Also, we do the
+        rem 32 or 64 bit detection here.
+        rem
+        gpc -v 2> temp
+        grep "gpc version 20070904" temp > temp2
+        if %errorlevel% neq 0 (
+        
+            echo *** Warning, Pascal-P5 is only validated to work with gpc version 20070904
+            
+        )
+        rm temp2
+        rem check 32/64 bit mode
+        set bits=64
+        grep "build=x86_64" temp > temp2
+        if %errorlevel% equ 0 (
+    
+            set bits=32
+    
+        )
+        
+    )
+    
+)
+
+rem
+rem Check all arguments. Note that we don't attempt to check or fix bad choices
+rem on the users part. We figure they know what they are doing.
 rem
 
 for %%x in (%*) do (
@@ -78,7 +123,7 @@ for %%x in (%*) do (
         echo "--32:        Select 32 bit target"
         echo "--64:        Select 64 bit target"
         echo
-        goto stop
+        exit /b 1
 
 	) else if "%%x" == "--gpc" (
 
@@ -99,7 +144,8 @@ for %%x in (%*) do (
     ) else (
     
         echo *** Option not recognized
-        goto stop
+        echo Terminating
+        exit /b 1
         
     )
 
@@ -115,72 +161,61 @@ rem
 rem GPC Pascal, named "gpc"
 rem
 
-if  not "%compiler%" == "ip_pascal" goto chkgpc
+if "%compiler%" == "ip_pascal" (
 
-rem
-rem Set up for IP Pascal
-rem
-cp ip_pascal\p5.bat      bin\p5.bat
-cp ip_pascal\compile.bat bin\compile.bat
-cp ip_pascal\run.bat     bin\run.bat
+    rem
+    rem Set up for IP Pascal
+    rem
+    cp ip_pascal\p5.bat      bin\p5.bat
+    cp ip_pascal\compile.bat bin\compile.bat
+    cp ip_pascal\run.bat     bin\run.bat
 
-cp ip_pascal\p5          bin\p5
-cp ip_pascal\compile     bin\compile
-cp ip_pascal\run         bin\run
+    cp ip_pascal\p5          bin\p5
+    cp ip_pascal\compile     bin\compile
+    cp ip_pascal\run         bin\run
 
-cp ip_pascal\Makefile    .
+    cp ip_pascal\Makefile    .
 
-cp ip_pascal\standard_tests/iso7185pat.cmp standard_tests
-cp ip_pascal\standard_tests/iso7185pats.cmp standard_tests
+    cp ip_pascal\standard_tests/iso7185pat.cmp standard_tests
+    cp ip_pascal\standard_tests/iso7185pats.cmp standard_tests
 
-rem
-rem IP Pascal does not care about line endings, but returning to DOS mode
-rem line endings normalizes the files for SVN checkin.
-rem
-rem doseol
+    rem
+    rem IP Pascal does not care about line endings, but returning to DOS mode
+    rem line endings normalizes the files for checkin.
+    rem
+    rem doseol
 
-echo Compiler set to IP Pascal
-goto setbits
+    echo Compiler set to IP Pascal
+    
+)
 
-:chkgpc
+if "%compiler%" == "gpc" (
 
-if  not "%compiler%" == "gpc" goto nonefound
+    rem
+    rem Set up for GPC Pascal
+    rem
+    cp gpc\p5.bat      bin\p5.bat
+    cp gpc\compile.bat bin\compile.bat
+    cp gpc\run.bat     bin\run.bat
 
-rem
-rem Set up for GPC Pascal
-rem
-cp gpc\p5.bat      bin\p5.bat
-cp gpc\compile.bat bin\compile.bat
-cp gpc\run.bat     bin\run.bat
+    cp gpc\p5          bin\p5
+    cp gpc\compile     bin\compile
+    cp gpc\run         bin\run
 
-cp gpc\p5          bin\p5
-cp gpc\compile     bin\compile
-cp gpc\run         bin\run
+    cp gpc/Makefile    .
 
-cp gpc/Makefile    .
+    cp gpc/standard_tests/iso7185pat.cmp standard_tests
+    cp gpc/standard_tests/iso7185pats.cmp standard_tests
 
-cp gpc/standard_tests/iso7185pat.cmp standard_tests
-cp gpc/standard_tests/iso7185pats.cmp standard_tests
+    rem
+    rem GPC needs Unix line endings in both the Unix and cygwin
+    rem versions.
+    rem
+    rem doseol
 
-rem
-rem GPC needs Unix line endings in both the Unix and cygwin
-rem versions.
-rem
-rem doseol
+    echo Compiler set to GPC Pascal
 
-echo Compiler set to GPC Pascal
-goto setbits
-
-rem
-rem No compiler name found!
-rem
-:nonefound
-echo *** Compiler name does not match currently implemented
-echo *** compilers.
-echo.
-echo IP Pascal  - use "ip_pascal"
-echo GPC Pascal - use "GPC"
-echo.
+)
 
 rem
 rem Set bit length
@@ -200,8 +235,3 @@ if "%bits%" == "64" (
 )
 
 echo Configure completed!
-
-rem
-rem Terminate script
-rem
-:stop
